#11 Feat: 게임결과 로딩화면
