+++ LunZh/LunZh/Sources/Feature/Group/Views/GroupMainView.swift
@@ -8,12 +8,7 @@
 import UIKit
 
 class GroupMainView: UIView {
-    
-    var isGameResultWaiting: Bool = false {
-        didSet {
-            updateMainBodyView()
-        }
-    }
+
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -33,7 +28,7 @@ class GroupMainView: UIView {
      public lazy var groupMainAppBar = GroupMainAppBar()
     
      // 그룹 메인화면 - 중앙 바디영역
-    public lazy var groupMainBody = UIView()
+    public lazy var groupMainBody = GroupMainBody()
     
     // MARK: - Function
     
@@ -53,21 +48,6 @@ class GroupMainView: UIView {
             $0.height.equalTo(60)
         }
         
-        groupMainBody.snp.makeConstraints {
-            $0.top.equalTo(groupMainAppBar.snp.bottom)
-            $0.horizontalEdges.bottom.equalTo(self)
-        }
-    }
-    
-    private func updateMainBodyView() {
-        // Remove existing body view from the view hierarchy
-        groupMainBody.removeFromSuperview()
-        
-        // Create and assign new body view based on the isGameResultWaiting state
-        groupMainBody = isGameResultWaiting ? GameResultWaitingView() : GroupMainBody()
-        
-        // Add the new body view to the view hierarchy and configure constraints
-        addSubview(groupMainBody)
         groupMainBody.snp.makeConstraints {
             $0.top.equalTo(groupMainAppBar.snp.bottom)
             $0.horizontalEdges.bottom.equalTo(self)
